requirements.txt
selenium==4.16.0
pytest==8.1.1
allure-pytest==2.13.5
Faker==25.1.0
python-dotenv==1.0.0